+++ jackz_vehicles.lua
@@ -2,9 +2,9 @@
 -- Created By Jackz
 -- SOURCE CODE: https://github.com/Jackzmc/lua-scripts
 local SCRIPT = "jackz_vehicles"
-local VERSION = "3.9.1"
+local VERSION = "3.9.2"
 local LANG_TARGET_VERSION = "1.3.3" -- Target version of translations.lua lib
-local VEHICLELIB_TARGET_VERSION = "1.1.5"
+local VEHICLELIB_TARGET_VERSION = "1.1.6"
 
 --#P:TEMPLATE("_SOURCE")
 --#P:TEMPLATE("common")